Move Backends.hs to Utils
