+++ persistent-migration.cabal
@@ -41,8 +41,8 @@ test-suite persistent-migration-test
   default-language:   Haskell2010
   main-is:            Unit.hs
   other-modules:      Test.Unit.Migration
-                      Test.Unit.Backends
                       Test.Unit.Validation
+                      Test.Unit.Utils.Backends
                       Test.Utils.Goldens
   build-depends: base
                , bytestring
@@ -63,9 +63,9 @@ test-suite persistent-migration-integration
   hs-source-dirs:     test
   default-language:   Haskell2010
   main-is:            Integration.hs
-  other-modules:      Test.Integration.Backends
-                      Test.Integration.Migration
+  other-modules:      Test.Integration.Migration
                       Test.Integration.Property
+                      Test.Integration.Utils.Backends
                       Test.Integration.Utils.RunSql
                       Test.Utils.Goldens
                       Test.Utils.QuickCheck
